+++ jannovar-filter/src/main/java/de/charite/compbio/jannovar/filter/WriterFilter.java
@@ -26,9 +26,9 @@ public class WriterFilter implements VariantContextFilter {
 	}
 
 	public void put(FlaggedVariant fv) throws FilterException {
-		LOGGER.trace("Variant added to writer {} => included? {}", new Object[] { fv.vc, fv.isIncluded() });
+		LOGGER.trace("Variant added to writer {} => included? {}", new Object[] { fv.getVC(), fv.isIncluded() });
 		if (fv.isIncluded())
-			writer.add(fv.vc);
+			writer.add(fv.getVC());
 	}
 
 	public void finish() {